pyproject.toml
[project]
name = "dump-truck"
version = "0.1.0"
description = "This is dumping ground!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.3.1",
    "bs4>=0.0.2",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-prebuilt>=1.0.5",
    "langgraph-sdk>=0.2.10",
    "langsmith>=0.4.48",
    "notebook>=7.5.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "trustcall>=0.0.39",
]
